requirements.txt
Flask==2.3.3
google-genai==1.55.0
requests==2.31.0
gunicorn==21.2.0
